fix 32 bit test error

+++ test/jld.jl
@@ -68,7 +68,7 @@ rng = 1:5
 objwithpointer = big(10)^10000
 # Custom BitsType (#99)
 bitstype 64 MyBT
-bt = reinterpret(MyBT, 55)
+bt = reinterpret(MyBT, int64(55))
 # Symbol arrays (#100)
 sa_asc = [:a, :b]
 sa_utf8 = [:α, :β]
